Fix method & property access when debase classes are extended.
When a class extends multiple debase classes, the instance’s
`interface` slot might not match up with the method / property name,
leading to noops.

In fact, it might be a terrible idea to include the interface name in
the class at all.  At the minimum, it should be a list of such, though
that seems like a major hassle.